move watchPoll to mixin
